+++ src/home/formats/hal.ts
@@ -1,27 +1,49 @@
+import { HalResource } from 'hal-types';
 import { getSetting } from '../../server-settings';
-import { HalBody } from '../../types';
+import { User } from '../../user/types';
 
-export default (version: string) => {
+export default (version: string, authenticatedUser: User) => {
 
-  const result: HalBody = {
+  const result: HalResource = {
     _links: {
       'self': { href: '/', title: 'Auth API Home' },
-      'sa:authorize' : { href: '/authorize', title: 'OAuth2 authorize endpoint', type: 'text/html' },
-      'sa:logout': { href: '/logout', title: 'Expire tokens and sessions' },
-      'sa:token': { href: '/token', title: 'OAuth2 protocol endpoint' },
-      'sa:validate-bearer': { href: '/validate-bearer', title: 'Validate a OAuth2 bearer token'},
-      'sa:validate-totp': { href: '/validate-totp', title: 'Validate a TOTP 2FA token + bearer token'},
-      'sa:user-collection': { href: '/user', title: 'List of users'},
-      'sa:change-password': { href: '/changepassword', title: 'Change user\'s password' },
-      'oauth_server_metadata_uri' : { href: '/.well-known/oauth-authorization-server', title: 'OAuth 2.0 Authorization Server Metadata' },
+      'authenticated-as': { href: '/user/' + authenticatedUser.id, title: authenticatedUser.nickname },
+      'authorize' : { href: '/authorize', title: 'OAuth2 authorize endpoint', type: 'text/html' },
+      'change-password': { href: '/changepassword', title: 'Change user\'s password' },
+      'introspect' : {
+        href: '/introspect',
+        title: 'OAuth2 Token Introspection',
+        hints: {
+          allow: ['POST'],
+        }
+      },
+      'logout': {
+        href: '/logout',
+        title: 'Log out'
+      },
+      'token': {
+        href: '/token',
+        title: 'OAuth2 protocol endpoint'
+      },
+      'validate-bearer': {
+        href: '/validate-bearer',
+        title: 'Validate a OAuth2 bearer token',
+        hints: {
+          status: 'deprecated',
+        }
+      },
+      'validate-totp': { href: '/validate-totp', title: 'Validate a TOTP 2FA token + bearer token'},
+      'user-collection': { href: '/user', title: 'List of users'},
+      'oauth_server_metadata_uri' : { href: '/.well-known/oauth-authorization-server', title: 'OAuth 2.0 Authorization Server Metadata' }
     },
     version: version,
   };
 
   if (getSetting('registration.enabled')) {
-    result._links['sa:registration'] = {
+    result._links.registration = {
       href: '/registration',
-      title: 'Create a new user account'
+      title: 'Create a new user account',
+      type: 'text/html'
     };
   }