+++ src/login/controller.ts
@@ -4,6 +4,7 @@ import { NotFound } from '@curveball/http-errors';
 import querystring from 'querystring';
 import log from '../log/service';
 import { EventType } from '../log/types';
+import { getSetting } from '../server-settings';
 import * as UserService from '../user/service';
 import { User } from '../user/types';
 import { loginForm } from './formats/html';
@@ -13,7 +14,10 @@ class LoginController extends Controller {
   async get(ctx: Context) {
 
     ctx.response.type = 'text/html';
-    ctx.response.body = loginForm(ctx.query.msg);
+    ctx.response.body = loginForm(
+      ctx.query.msg,
+      await getSetting('registration.enabled')
+    );
 
   }