+++ src/routes.ts
@@ -3,6 +3,7 @@ import blob from './blob/controller';
 import changePassword from './changepassword/controller';
 import health from './health/controller';
 import home from './home/controller';
+import introspect from './introspect/controller';
 import userLog from './log/controller/user';
 import login from './login/controller';
 import logout from './logout/controller';
@@ -23,6 +24,7 @@ const routes = [
   router('/login', login),
   router('/assets/:filename', blob),
   router('/health', health),
+  router('/introspect', introspect),
   router('/register', register),
   router('/authorize', oauth2Authorize),
   router('/validate-bearer', validateBearer),